[ISSUE-7] Organized files under /Model
