+++ src/Model/Authentication/Providers.ts
@@ -1,0 +1,4 @@
+export enum Providers {
+    GOOGLE = 'google',
+    FACEBOOK = 'facebook',
+}
