+++ src/Model/Authentication/SignInTypes.ts
@@ -1,0 +1,4 @@
+export enum SignInTypes {
+    POP_UP = 'popup',
+    REDIRECT = 'redirect',
+}
